+++ client/src/pages/homepage/featuredRestaurants.jsx
@@ -1,25 +1,12 @@
 import React, { useState,useEffect } from "react";
-import { getRestaurants } from "../../services/location.services";
 
 
 
 
-const FeaturedRestaurants= () => {
 
-  const [restaurants,setRestaurants] = useState([]);
-  useEffect(() => {
-  const fetch_restauarnts = async () => {
-    
-      const result = await getRestaurants();
-      console.log(result)
-      setRestaurants(result);
+const FeaturedRestaurants= ({restaurants}) => {
 
-   
-  }
-
-  fetch_restauarnts();
-},[]);
-
+  
   
   return (
     <section className="my-12">
@@ -28,7 +15,7 @@ const FeaturedRestaurants= () => {
       <p className="text-gray-600 mb-6">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras iaculis consectetur nisi sagittis.</p>
       
       <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-        {restaurants.map((restaurant, index) => (
+        {restaurants && restaurants.length > 0 ? restaurants.map((restaurant, index) => (
           <div key={`restaurant-${restaurant.id}`} className="bg-white rounded-lg shadow-lg p-4">
             <img src={restaurant.cover_image_url} alt={restaurant.description} className="w-full h-48 object-cover rounded-lg" />
             <div className="flex justify-between items-center mt-3">
@@ -41,7 +28,7 @@ const FeaturedRestaurants= () => {
             <p className="text-gray-600 text-sm mt-2">{restaurant.description}</p>
             
           </div>
-        ))}
+        )): "No Data"}
       </div>
     
     </section>
